% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Aneuploid_table_fitted_3day_210212}
\alias{Aneuploid_table_fitted_3day_210212}
\title{Aneuploidy data
Contains maternal age and probability of aneuploidy.}
\format{
A data frame with 28 rows and 2 variables:
 \describe{
   \item{age}{maternal age}
   \item{prob}{probability of aneuploidy}
   }
 @source {Derived from publication}
 @examples
 data(Aneuploid_table_fitted_3day_210212)
}
\usage{
Aneuploid_table_fitted_3day_210212
}
\description{
Aneuploidy data
Contains maternal age and probability of aneuploidy.
}
\keyword{datasets}
