% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aroonda-data.R
\docType{data}
\name{Aroonda}
\alias{Aroonda}
\title{Aroonda simulation conditions}
\format{
An object of class \code{sim_conds}; see \linkS4class{sim_conds}.
}
\usage{
data(Aroonda)
}
\description{
A sim_conds object defining the simulation conditions
for Aroonda during the growing season of 2022.
Can be used to run the endosymbiont model.
Environmental data derived from SILO.
Simulation set to run from 2022-04-01 to 2022-11-01.\cr
}
\keyword{datasets}
