% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_thresh.R
\name{auc_thresh}
\alias{auc_thresh}
\title{Area between thresholds}
\usage{
auc_thresh(x, y, min_val, max_val)
}
\arguments{
\item{x}{numeric vector of x values}

\item{y}{numeric vector of y values}

\item{min_val}{lower threshold of y}

\item{max_val}{upper threshold of y}
}
\value{
a vector consisting of two arguments: the full area under the curve, and the area under the curve between the threshold values
}
\description{
Calculate the area of a curve between thresholds of y values
}
\details{
This function uses smooth splines to fit a curve defined by vectors of x and y values. It then integrates across the curve to find the area under the curve, and between two threshold values of y.
}
\examples{
temp_auc <- EndosymbiontModel:::auc_thresh(seq(0,50), sin(seq(0,50)), 0, 0.5)
temp_auc
}
\keyword{internal}
