% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_species.R
\name{check_species}
\alias{check_species}
\title{Check species object}
\usage{
check_species(object)
}
\arguments{
\item{object}{An object of class \code{species}; see \link{species}}
}
\description{
Checks the validity of a species object
}
