% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterfact.R
\name{counterfact}
\alias{counterfact}
\title{Counterfactual Simulations}
\usage{
counterfact(
  Pest,
  Endosymbiont,
  Crop,
  Parasitoid = NULL,
  init,
  conds,
  modules = NULL
)
}
\arguments{
\item{Pest}{object of class \code{pest} defining the pest}

\item{Endosymbiont}{object of class \code{endosym} defining the endosymbiont}

\item{Crop}{object of class \code{crop} defining the crop}

\item{Parasitoid}{object of class \code{parasitoid} defining the parasitoid}

\item{init}{object of class \code{initial} defining the starting conditions for the simulation}

\item{conds}{object of class \code{sim_conds} defining the simulation conditions}

\item{modules}{list of modules and scenarios to run, if \code{NULL} (default) all fixed to FALSE}
}
\value{
object of class \code{endosim_col}; see \linkS4class{endosim_col}.
}
\description{
Run multiple simulations of the endosymbiont model under different scenarios
}
\details{
Uses the \link{endosim} function to run simulations of the endosymbiont model. A basic null model is always run, without any modules. The provided \linkS4class{pest}, \linkS4class{endosym}, \linkS4class{crop},\linkS4class{parasitoid}, \linkS4class{initial}, and \linkS4class{sim_conds} objects are used to parameterise the simulation.

\code{modules} should be a list with the following arguments: \code{vert_trans}, \code{hori_trans}, \code{imi}, \code{emi}, \code{para}. Each argument should be a logical vector, consisting either of \code{FALSE} (fixed off), \code{TRUE} (fixed on), or \code{c(FALSE, TRUE)} (run counterfactual scenarios with module turned on and turned off).
}
\seealso{
\code{\link[=endosim]{endosim()}}
}
