% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\docType{class}
\name{crop-class}
\alias{crop-class}
\title{An S4 class to store a crop's life history for the endosymbiont model}
\description{
An S4 class to store a crop's life history for the endosymbiont model
}
\section{Slots}{

\describe{
\item{\code{name}}{The name of the crop}

\item{\code{heal_time}}{Half-time in days for plant recovery from endosymbiont infection}

\item{\code{sowing_date}}{Date crop is sown in YYYY-MM-DD format}

\item{\code{harvest_date}}{Date crop is harvested in YYYY-MM-DD format}
}}

\examples{
canola <-
new("crop",
name = "Canola",
heal_time = 2,
sowing_date = "2022-04-01",
harvest_date = "2022-11-01")

canola
}
