% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{cycleToOutcome}
\alias{cycleToOutcome}
\title{Tabulate number of cycles to future outcome in an EndoSim simulation object}
\usage{
cycleToOutcome(obj)
}
\arguments{
\item{obj}{An outcome slot from simulation object (required).}
}
\value{
Table of number of cycles to next outcome
}
\description{
This function tabulates the number of cycles to future outcome for any given cycle in an EndoSim simulation object.
}
