% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_silo.R
\name{download_silo}
\alias{download_silo}
\title{Download SILO data}
\usage{
download_silo(date, year, var, path)
}
\arguments{
\item{date}{date to download data for in YYYY-MM-DD format}

\item{year}{year of date}

\item{var}{name of variable to download (e.g. max_temp, min_temp, daily_rain)}

\item{path}{pathway to output directory}
}
\value{
raster of selected weather data variable for selected date
}
\description{
Download weather station data from SILO
}
\seealso{
\code{\link[=make_conds]{make_conds()}}
}
