% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endosim_col.R
\docType{class}
\name{endosim_col-class}
\alias{endosim_col-class}
\title{An S4 class to store a collection of simulations with the endosymbiont model}
\description{
An S4 class to store a collection of simulations with the endosymbiont model
}
\section{Slots}{

\describe{
\item{\code{scenarios}}{Dataframe defining which modules are active or deactivated in each simulated scenario}

\item{\code{sims}}{List of objects of class \linkS4class{endosim_mod}, each corresponding to a row of \code{scenarios}}
}}

