% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endosym.R
\docType{class}
\name{endosym-class}
\alias{endosym-class}
\alias{endosym}
\title{An S4 class to store a endosymbiont's life history for the endosymbiont model}
\description{
An S4 class to store a endosymbiont's life history for the endosymbiont model
}
\section{Slots}{

\describe{
\item{\code{name}}{The name of the endosymbiont}

\item{\code{fitness_cost}}{Relative fitness cost of R+ phenotype (0-1)}

\item{\code{fun_trans_eff}}{Function describing temperature-dependent efficiency of endosymbiont transmission to plants}

\item{\code{fun_susc}}{Function describing temperature-dependent susceptibility of R- pests to infection from plants}
}}

\examples{
Rickettsiella <-
new("endosym",
name = "Rickettsiella",
fitness_cost = 0.5,
fun_trans_eff = EndosymbiontModel:::fit_custom(seq(0, 36, 4), c(0, 0.6, 0.7, 0.75, 0.8, 0.8, 0.75, 0.6, 0.3, 0), k = 4),
fun_susc = EndosymbiontModel:::fit_custom(seq(0, 40, 4), c(0, 0.15, 0.3, 0.45, 0.6, 0.75, 0.75, 0.6, 0.3, 0, 0), k = 7)
)

Rickettsiella
}
