% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endosym_col.R
\docType{class}
\name{endosym_col-class}
\alias{endosym_col-class}
\title{An S4 class to store a collection of simulations with the Endosymbiont model}
\description{
An S4 class to store a collection of simulations with the Endosymbiont model
}
\section{Slots}{

\describe{
\item{\code{scenarios}}{Dataframe defining which modules are active or deactivated in each simulated scenario}

\item{\code{sims}}{List of objects of class \linkS4class{endosym_mod}, each corresponding to a row of \code{scenarios}}
}}

