% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endosym_mod.R
\docType{class}
\name{endosym_mod-class}
\alias{endosym_mod-class}
\title{An S4 class to store a simulation output}
\description{
An S4 class to store a simulation output
}
\section{Slots}{

\describe{
\item{\code{pest}}{name of the pest}

\item{\code{crop}}{name of the crop}

\item{\code{endosymbiont}}{name of the endosymbiont}

\item{\code{parasitoid}}{name of the parasitoid}

\item{\code{start_date}}{first day of simulation in YYYY-MM-DD format}

\item{\code{sim_length}}{length of simulation in days}

\item{\code{vert_trans}}{is vertical transmission active}

\item{\code{hori_trans}}{is horizontal transmission active}

\item{\code{imi}}{is immigration active}

\item{\code{emi}}{is emigration active}

\item{\code{para}}{is parasitoid active}

\item{\code{pest_df}}{data.frame of pest population through time}

\item{\code{pest_cohorts}}{array of pest cohorts at the end of the simulation}

\item{\code{para_df}}{data.frame of parasitoid population through time}

\item{\code{area}}{area of crop (m2)}
}}

