% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endosym_model.R
\name{endosym_model}
\alias{endosym_model}
\title{Endosymbiont model}
\usage{
endosym_model(
  Pest,
  Crop,
  Beneficials = NULL,
  init,
  conds,
  plot = TRUE,
  vert_trans = TRUE,
  hori_trans = TRUE,
  imi = TRUE,
  emi = TRUE
)
}
\arguments{
\item{Pest}{object of class \code{pest} defining the pest}

\item{Crop}{object of class \code{crop} defining the crop}

\item{Beneficials}{list of objects of class \code{beneficial} describing the beneficials in the system}

\item{init}{object of class \code{initial} defining the starting conditions for the simulation}

\item{conds}{object of class \code{sim_conds} defining the simulation conditions}

\item{plot}{if \code{TRUE} (default) generate plots}

\item{vert_trans}{set to \code{FALSE} to turn off vertical transmission}

\item{hori_trans}{set to \code{FALSE} to turn off horizontal transmission}

\item{imi}{set to \code{FALSE} to turn off immigration into paddock}

\item{emi}{set to \code{FALSE} to turn off emigration from paddock}
}
\value{
object of class \code{endosym_mod}; see \link{endosym_mod}.
}
\description{
Run simulation using the endosymbiont model
}
\details{
Constructs the endosymbiont model using the provided \link{pest}, \link{crop}, \link{beneficials}, \link{initial}, and \link{sim_conds} objects.

The \code{Pest} species feeds on the provided \code{Crop}.

Vertical transmission of endosymbiont is defined using functions in the \code{Pest}. Horizontal transmission through herbivory is defined using functions in the \code{Crop}.

Each element of the \code{Beneficials} list should be a \code{beneficial} object with functions defining the interactions with the pest (e.g. predation, parasitism) and their impacts on mortality and endosymbiont transmission.
}
