% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_alate.R
\name{fit_alate}
\alias{fit_alate}
\title{Fit alate production function}
\usage{
fit_alate(omega, z, W, V)
}
\arguments{
\item{omega}{inflection point of logistic regression}

\item{z}{logistic regression gradient parameter}

\item{W}{logistic regression gradient parameter}

\item{V}{logistic regression gradient parameter}
}
\value{
function that accepts x and returns y based on fitted model
}
\description{
Fits a function that returns proportion of offspring destined to develop into alates
}
\details{
Fits a logistic regression curve to determine the proportion of offspring destined to become alate adults, based on a function defined for cherry-oat aphids (\emph{Rhopalosiphum padi}; Morgan 2000; Parry et al. 2006)
\deqn{
\frac{z}{1 + e^{-W \cdot \left( \frac{\text{density}}{V} - \omega \right)}}
}
}
\examples{
my_func <- EndosymbiontModel:::fit_alate(67.418, 0.993, 0.076, 300)
plot(seq(0, 15000, by = 1), my_func(seq(0, 15000, by = 1)), type = "l")
}
\references{
Morgan, D., Population dynamics of the bird cherry-oat aphid, \emph{Rhopalosiphum padi} (L.), during the autumn and winter: a modelling approach. Agricultural and Forest Entomology, 2, 297-304 (2000)
Parry, H.R., Evans, A.J., Morgan, D., Aphid population response to agricultural landscape change: a spatially explicit, individual-based model. Ecological Modelling, 199, 451-463 (2006)
}
\seealso{
\code{\link[=fit_null]{fit_null()}}
}
\keyword{internal}
