% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_bannerman.R
\name{fit_bannerman}
\alias{fit_bannerman}
\title{Fit Bannerman & Roitberg function for fecundity}
\usage{
fit_bannerman(D, p)
}
\arguments{
\item{D}{carrying capacity (inflection point)}

\item{p}{rate of decline in fecundity}
}
\value{
function that accepts x and returns y based on fitted model
}
\description{
Fits a function that returns scaled daily fecundity based on population density
}
\details{
Fits a curve based on Bannerman & Roitberg (2014) to describe how daily fecundity changes with density:
\deqn{
1 - \frac{1}{1 + e^{-p \cdot (\text{density} - D)}}
}
}
\examples{
my_func <- EndoSim:::fit_bannerman(10000, 0.0008)
plot(seq(0, 15000, by = 1), my_func(seq(0, 15000, by = 1)), type = "l")
}
\references{
Bannerman, J,A., Roitberg, B.D., Impact of extreme and fluctuating temperatures on aphid-parasitoid dynamics. Oikos, 123, 89-98 (2014)
}
\seealso{
\code{\link[=fit_null]{fit_null()}}, \code{\link[=fit_quadratic]{fit_quadratic()}}, \code{\link[=fit_weibull]{fit_weibull()}}
}
\keyword{internal}
