% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_briere.R
\name{fit_briere}
\alias{fit_briere}
\title{Fit Brière curve}
\usage{
fit_briere(ymax, xopt, xmin, xmax)
}
\arguments{
\item{ymax}{maximal value of y}

\item{xopt}{value of x at which y = ymax}

\item{xmin}{minimal threshold of x (y = 0)}

\item{xmax}{maximal threshold of x (y = 0)}
}
\value{
function that accepts x and returns y based on fitted model
}
\description{
Fits a curve using the Brière model for fitting thermal performance curves
}
\details{
Fits a curve using the Brière model (Brière et al. 1999):
\deqn{
  y = a \cdot x \cdot (x - x_{\text{min}}) \cdot ((x_{\text{max}} - x)^{\frac{1}{b}})
}
Optimal values for \code{a} (scale parameter) and \code{b} (shape parameter) are estimated based on provided arguments.
}
\references{
Brière, J.F., Pracros, P., Le Roux, A.Y., Pierre, J.S., A novel rate model of temperature-dependent development for arthropods. Environmental Entomololgy, 28, 22–29 (1999)
}
\keyword{internal}
