% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_emi_thack.R
\name{fit_emi_thack}
\alias{fit_emi_thack}
\title{Fit Thackray model for emigration}
\usage{
fit_emi_thack()
}
\arguments{
\item{x}{population of newly emerged adults}
}
\value{
number of emigrating adults
}
\description{
Fits a function that returns number of emigrating adults
}
\details{
A base emigration rate of 20\% of newly emerged adults
}
\examples{
my_func <- EndoSim:::fit_emi_thack()
my_func(runif(10, 1, 100))
}
\references{
Thackray, D.J., Diggle, A.J., Berlandier, F.A., Jones, R.A.C., Forecasting aphid outbreaks and epidemics of Cucumber mosaic virus in lupin crops in a Mediterranean-type environment. Virus Research, 100, 67-82 (2004)
}
\seealso{
\code{\link[=fit_null]{fit_null()}}
}
\keyword{internal}
