% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gaussian.R
\name{fit_gaussian}
\alias{fit_gaussian}
\title{Fit Gaussian curve}
\usage{
fit_gaussian(ymax, xopt, xmin, xmax)
}
\arguments{
\item{ymax}{maximal value of y}

\item{xopt}{value of x at which y = ymax}

\item{xmin}{minimal threshold of x (y = 0)}

\item{xmax}{maximal threshold of x (y = 0)}
}
\value{
function that accepts x and returns y based on fitted model
}
\description{
Fits a curve using the Gaussian model for fitting thermal performance curves
}
\details{
Fits a curve using the Gaussian model:
\deqn{
  y = y_{\text{max}} \cdot e^{-0.5 \cdot \left(\frac{{|x - x_{\text{opt}}|}}{{a}}\right)^2}
}
Optimal value for \code{a} is estimated based on provided arguments.
}
\keyword{internal}
