% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gompertz.R
\name{fit_gompertz}
\alias{fit_gompertz}
\title{Fit Gompertz survival function}
\usage{
fit_gompertz(delta, beta)
}
\arguments{
\item{delta}{shape of mortality curve}

\item{beta}{rate of decline}
}
\value{
function that accepts x and returns y based on fitted model
}
\description{
Fits a function that returns mortality rate based on an inverse Gompertz survival function
}
\details{
A mortality rate function to pass on to the Endosymbiont model.
Fits an a Gomeprtz survival function to calculate mortality rate:
\deqn{
1 - e^{\left(\frac{\delta}{\beta}\left(1 - e^{\beta \cdot \text{{age}}}\right)\right)}
}
}
\examples{
my_func <- EndosymbiontModel:::fit_gompertz(0.00188, 0.13)
plot(seq(0, 50, by = 0.1), my_func(seq(0, 50, by = 0.1)), type = "l")
}
\references{
Gompertz, B., On the nature of the function expressive of the law of human mortality and on a new mode of determining the value of life contingencies. Philosophical Transactions of the Royal Society of London, 115, 513-583 (1825)
}
\seealso{
\code{\link[=fit_null]{fit_null()}}, \code{\link[=fit_rainfall_thack]{fit_rainfall_thack()}}
}
\keyword{internal}
