% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lactin.R
\name{fit_lactin}
\alias{fit_lactin}
\title{Fit Lactin model for development}
\usage{
fit_lactin(epsilon, tau, lambda, Tupper)
}
\arguments{
\item{epsilon}{development rate scalar}

\item{tau}{development rate scalar}

\item{lambda}{development rate scalar}

\item{Tupper}{upper development threshold}
}
\value{
function that accepts x and returns y based on fitted model
}
\description{
Fits a function that returns parasitoid wasp development rate based on daily temperature
}
\details{
A development rate function to pass on to the Parasitoid model.
Fits a curve using the Lactin model (Lactin et al. 1995):
\deqn{
e^{\epsilon \cdot T} - e^{\epsilon \cdot T_{\text{opt}} - \left(\frac{T_{\text{opt}} - T}{\tau}\right)} + \lambda
}
}
\examples{
my_func <- EndosymbiontModel:::fit_lactin(0.13048, 7.64313, -0.01308, 35.47)
plot(seq(0, 50, by = 0.1), my_func(seq(0, 50, by = 0.1)), type = "l")
}
\references{
Lactin, D.J., Holliday, N.J., Johnson, D.L., Craigen, R., Improved rate model of temperature-dependent development by arthropods. Environmental Entomology, 24, 68-75 (1995)
}
\seealso{
\code{\link[=fit_null]{fit_null()}}, \code{\link[=fit_wang]{fit_wang()}}
}
\keyword{internal}
