% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_null.R
\name{fit_null}
\alias{fit_null}
\title{Fit null}
\usage{
fit_null(output = 0)
}
\arguments{
\item{output}{which value to return (default is 0)}

\item{x}{vector of any class}
}
\value{
function that accepts x and returns the value of \code{output}
}
\description{
Fits a function that returns 0 for any given input value
}
\details{
A function to pass on to the Endosymbiont model which can be used to nullify any process. The default output is 0, and therefore can be used to stand-in for any process that is modelled as a function accepting an input vector. However, the output can be changed to a different value (e.g., 1) if required, as long as the null process is supposed to return a constant value.
}
\examples{
my_func <- EndosymbiontModel:::fit_null(0)
plot(seq(0, 50, by = 0.1), my_func(seq(0, 50, by = 0.1)), type = "l")

my_func <- EndosymbiontModel:::fit_null(1)
plot(seq(0, 50, by = 0.1), my_func(seq(0, 50, by = 0.1)), type = "l")

my_func <- EndosymbiontModel:::fit_null(0.5)
plot(seq(0, 50, by = 0.1), my_func(seq(0, 50, by = 0.1)), type = "l")
}
\seealso{
\code{\link[=fit_custom]{fit_custom()}}
}
\keyword{internal}
