% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_rainfall_thack.R
\name{fit_rainfall_thack}
\alias{fit_rainfall_thack}
\title{Fit Thackray model for rainfall-induced mortality}
\usage{
fit_rainfall_thack()
}
\arguments{
\item{x}{daily rainfall in mm}
}
\value{
mortality rate
}
\description{
Fits a function that returns mortality rate based on daily rainfall
}
\details{
A mortality rate function to pass on to the Endosymbiont model.
Fits a curve using the Thackray definition of rainfall-induced mortality in aphids (Thackray et al. 2004, modified for green peach aphid by Maling et al. 2010):
\deqn{
\text{{Rainfall Mortality}} = 0.026 \cdot \text{{Rainfall}} - 0.33
}

Where mortality increases linearly as daily rainfall increases from 15 mm and is capped at a maximum of 0.5.
}
\examples{
my_func <- EndoSim:::fit_rainfall_thack()
plot(seq(0, 50, by = 0.1), my_func(seq(0, 50, by = 0.1)), type = "l")
}
\references{
Thackray, D.J., Diggle, A.J., Berlandier, F.A., Jones, R.A.C., Forecasting aphid outbreaks and epidemics of Cucumber mosaic virus in lupin crops in a Mediterranean-type environment. Virus Research, 100, 67-82 (2004)

Maling, T., Diggle, A.J., Thackray, D.J., Siddique, K.H.M., Jones, R.A.C. An epidemiological model for externally acquired vector-borne viruses applied to \emph{Beet western yellows virus} in \emph{Brassica napus} crops in a Mediterranean-type environment. Crop and Pasture Science, 61, 132-144 (2010)
}
\seealso{
\code{\link[=fit_gompertz]{fit_gompertz()}}, \code{\link[=fit_null]{fit_null()}}
}
\keyword{internal}
