% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_rezende.R
\name{fit_rezende}
\alias{fit_rezende}
\title{Fit Rezende curve}
\usage{
fit_rezende(ymax, xopt, xmin, xmax)
}
\arguments{
\item{ymax}{maximal value of y}

\item{xopt}{value of x at which y = ymax}

\item{xmin}{minimal threshold of x (y = 0)}

\item{xmax}{maximal threshold of x (y = 0)}
}
\value{
function that accepts x and returns y based on fitted model
}
\description{
Fits a curve using the Rezende model for fitting thermal performance curves
}
\details{
Fits a curve using the Rezende model (Rezende et al. 2019):
\deqn{
  y = \begin{cases}
  a \cdot 10^{\frac{\log_{10}(q_{10})}{\frac{10}{x}}}, & \text{if } x < x_{\text{opt}} \\
  a \cdot 10^{\frac{\log_{10}(q_{10})}{\frac{10}{x}}} \cdot (1 - b \cdot (x_{\text{opt}} - x)^2), & \text{otherwise}
  \end{cases}
}
Optimal values for \code{a} (rate shift parameter), \code{b} (decline rate parameter), and q_{10} are estimated based on provided arguments.
}
\examples{
my_func <- EndoSim:::fit_rezende(1, 25, 2.84, 32.79)
plot(seq(0, 50, by = 0.1), my_func(seq(0, 50, by = 0.1)), type = "l")
}
\references{
Rezende, Enrico L., and Francisco Bozinovic. Thermal performance across levels of biological organization. Philosophical Transactions of the Royal Society B 374.1778 (2019): 20180549.
}
\seealso{
\code{\link[=fit_custom]{fit_custom()}}, \code{\link[=fit_gaussian]{fit_gaussian()}}, \code{\link[=fit_null]{fit_null()}}, \code{\link[=fit_quadratic]{fit_quadratic()}}
}
\keyword{internal}
