% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sigmoid.R
\name{fit_sigmoid}
\alias{fit_sigmoid}
\title{Fit sigmoid curve}
\usage{
fit_sigmoid(ymax, xmin, xmax)
}
\arguments{
\item{ymax}{maximal value of y}

\item{xmin}{minimal threshold of x (y = 0)}

\item{xmax}{maximal threshold of x (y = ymax)}
}
\value{
function that accepts x and returns y based on fitted model
}
\description{
Fits a sigmoid curve
}
\details{
Fits a sigmoid curve:
\deqn{
  y = \frac{a}{1 + e^{-b \cdot (x - x_{\text{max}})}}
}
Optimal values for \code{a} and \code{b} (curve breadth parameter) are estimated based on provided arguments.
}
\keyword{internal}
