% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_supergaus.R
\name{fit_supergaus}
\alias{fit_supergaus}
\title{Fit super-gaussian function for mortality}
\usage{
fit_supergaus(Tupper, Tlower, Piupper, Pilower, Psi, Chi)
}
\arguments{
\item{Tupper}{upper development threshold}

\item{Tlower}{lower development threshold}

\item{Piupper}{mortality rate beyond upper threshold}

\item{Pilower}{mortality rate beyond lower threshold}

\item{Psi}{survival breadth parameter}

\item{Chi}{survival breadth parameter}
}
\value{
function that accepts Tmax and Tmin and returns y based on fitted model
}
\description{
Fits a function that returns aphid mortality rate based on divergence of maximum and minimum temperatures from development thresholds
}
\details{
A temperature-dependent mortality rate function to pass on to the Endosymbiont model.
Fits a super-gaussian function:
\deqn{
  1 - e^ ( -\Pi_{\text{upper}} \cdot (\chi \cdot (\frac{T_{\text{max}} - T_{\text{upper}}}{\Psi})^2) -\Pi_{\text{lower}} \cdot (\chi \cdot (\frac{T_{\text{lower}} - T_{\text{min}}}{\Psi})^2) )
}
}
\examples{
my_func <- EndoSim:::fit_supergaus(32.79, 2.84, 0.00083, 0.00017, 7.66, 17.66)
plot(seq(-10, 50, by = 0.1), sapply(seq(-10, 50, by = 0.1), function(x) my_func(x, x)), type = "l")
}
\seealso{
\code{\link[=fit_null]{fit_null()}}
}
\keyword{internal}
