% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_thirdpoly.R
\name{fit_thirdpoly}
\alias{fit_thirdpoly}
\title{Fit third-order polynomial regression}
\usage{
fit_thirdpoly(a, b, c, d, inv = F)
}
\arguments{
\item{a}{coefficient of the cubic term (x^3)}

\item{b}{coefficient of the quadratic term (x^2)}

\item{c}{coefficient of the linear term (x)}

\item{d}{constant term}

\item{inv}{if \code{TRUE} return inverse}
}
\value{
function that accepts x and returns y based on fitted model
}
\description{
Fits a function that returns a third-order polynomial regression
}
\details{
Fits a third-order polynomial regression:
\deqn{
a \cdot x^3 + b \cdot x^2 + c \cdot x + d
}
}
\examples{
my_func <- EndoSim:::fit_thirdpoly(0.00001084, -0.0021, -0.0272, -0.0145)
plot(seq(0, 50, by = 0.1), my_func(seq(0, 50, by = 0.1)), type = "l")
}
\seealso{
\code{\link[=fit_null]{fit_null()}}
}
\keyword{internal}
