% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tpc.R
\name{fit_tpc}
\alias{fit_tpc}
\title{Fit thermal performance curve}
\usage{
fit_tpc(model, ...)
}
\arguments{
\item{model}{one of "briere", "custom", "gaussian", "quadratic", "rezende", "weibull"}

\item{...}{additional arguments of selected model}
}
\value{
function that accepts x and returns y based on fitted model
}
\description{
Fits a thermal performance curve using selected model
}
\details{
Fits a curve using the selected model.
If "custom" requires vectors of x and y values.
For all other models requires values of ymax, xopt, xmin, and xmax and optimises parameters of selected model based on input arguments.
}
\examples{
my_func1 <- EndosymbiontModel:::fit_tpc("briere", 0.16, 22.61, 4, 30)
plot(seq(0, 50, by = 0.1), my_func1(seq(0, 50, by = 0.1)), type = "l")

my_func2 <- EndosymbiontModel:::fit_tpc("custom", c(4, 8.87, 11.82, 14.22, 17.13, 19.92, 22.61, 25.77, 27.80, 30), c(0, 0.06, 0.09, 0.11, 0.13, 0.15, 0.16, 0.14, 0.04, 0))
plot(seq(0, 50, by = 0.1), my_func2(seq(0, 50, by = 0.1)), type = "l")
}
\seealso{
\code{\link[=fit_briere]{fit_briere()}}, \code{\link[=fit_custom]{fit_custom()}}, \code{\link[=fit_gaussian]{fit_gaussian()}}, \code{\link[=fit_quadratic]{fit_quadratic()}}, \code{\link[=fit_rezende]{fit_rezende()}}, \code{\link[=fit_weibull]{fit_weibull()}}
}
\keyword{internal}
