% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{futureEndoSim}
\alias{futureEndoSim}
\title{Tabulate future outcome in an EndoSim simulation object}
\usage{
futureEndoSim(sim)
}
\arguments{
\item{sim}{An outcome slot from simulation object (required).}
}
\value{
Table of future outcomes
}
\description{
This function tabulates the future outcome for any given cycle in an EndoSim simulation object.
}
