% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPA-data.R
\docType{data}
\name{GPA}
\alias{GPA}
\title{GPA species information}
\format{
An object of class \code{species}; see \link{species}.
}
\usage{
data(GPA)
}
\description{
A species object defining the life history and function
for green peach aphid (\emph{Myzus persicae}).
Can be used to run the endosymbiont model on GPA.\cr
\cr
\code{f_growth} fitted using \link{fit_custom} with data from Whalon & Smilowitz 1979.\cr
\code{f_loss} fitted using \link{fit_sigmoid} with data from Maling et al. 2008.\cr
\code{f_suscept} fitted using \link{fit_custom} with data from Thackray et al. 2009.\cr
\code{f_inoc} fitted using \link{fit_quadratic} with data from Thackray et al. 2009.\cr
\code{f_alate} fitted using \link{fit_custom} with data from Thackray et al. 2009.\cr
}
\references{
Maling, T., Diggle, A.J., Thackray, D.J., Siddique, K.H.M., Jones, R.A.C., An epidemiological model for externally sourced vector-borne viruses applied to \emph{Bean yellow mosaic virus} in lupin crops in a Mediterranean-type environment. Phytopathology, 98, 1280–1290 (2008)

Thackray, D.J., Diggle, A.J., Jones, R.A.C., BYDV PREDICTOR: a simulation model to predict aphid arrival, epidemics of \emph{Barley yellow dwarf virus} and yield loss in wheat crops in a Mediterranean-type environment. Plant Pathology, 58, 186–202 (2009)]

Whalon, M.E., Smilowitz, Z., Temperature-dependent model for predicting field populations of green peach aphid, \emph{Myzus persicae}. The Canadian Entomologist, 111, 1025–1032 (1979)
}
\keyword{datasets}
