% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial.R
\docType{class}
\name{initial-class}
\alias{initial-class}
\title{An S4 class to store initial population sizes for the endosymbiont model}
\description{
An S4 class to store initial population sizes for the endosymbiont model
}
\section{Slots}{

\describe{
\item{\code{Pest}}{Array describing sizes of pest cohorts of each developmental stage, divided into R+ and R- apterae and alates}

\item{\code{Crop}}{Dataframe where each row is an individual plant, the first column (t) describes the time of infection, the second and third columns binary representations of infection status - if R+ 1 in second column and 0 in third, if R- 0 in second column and 1 in third}

\item{\code{Parasitoid}}{Numeric describing how many adult female parasitoids at start of simulation}
}}

