% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_conds.R
\name{make_conds}
\alias{make_conds}
\title{Create sim_conds}
\usage{
make_conds(start_date, end_date, lat, long, path = NULL)
}
\arguments{
\item{start_date}{start date of simulation in YYYY-MM-DD format}

\item{end_date}{end date of simulation in YYYY-MM-DD format}

\item{lat}{latitude of site in decimal degrees}

\item{long}{longitude of site in decimal degrees}

\item{path}{pathway to SILO weather data. If null (default) data for the selected dates are downloaded}
}
\value{
object of class \code{sim_conds} with the following arguments:
\itemize{
\item env: dataframe containing mean temperature and rainfall for each day of the simulation
\item sim_length: length of simulation in days
\item start_date: start date of simulation in YYYY-MM-DD format
}
}
\description{
Create sim_conds object with conditions for simulation
}
\seealso{
\code{\link[=download_silo]{download_silo()}}
}
