% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{pairedSampling}
\alias{pairedSampling}
\title{Summarise paired samples in an EndoSim simulation object}
\usage{
pairedSampling(sim_object, sampling_object, sep_cycles = 1)
}
\arguments{
\item{sim_object}{An outcome slot from simulation object (required).}

\item{sampling_object}{A sampling object (required).}

\item{sep_cycles}{Number of cycles between samplings. Default value 1.}
}
\value{
Expanded sampling object with paired data.
}
\description{
This function created paired sample data from EndoSim simulation object.
}
