% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parasitoid.R
\docType{class}
\name{parasitoid-class}
\alias{parasitoid-class}
\title{An S4 class to store a parasitoid's life history for the endosymbiont model}
\description{
An S4 class to store a parasitoid's life history for the endosymbiont model
}
\section{Slots}{

\describe{
\item{\code{species}}{The name of the pest}

\item{\code{fun_dev_para}}{Function describing temperature-dependent development of parasitoids}

\item{\code{fun_para_scal}}{Function to modify development rate}

\item{\code{fun_attack}}{Function describing attack rate as function of temperature}

\item{\code{fun_handling}}{Function describing handling rate as function of temperature}

\item{\code{susc_stage}}{Numeric vector defining the susceptible lifestages of the pest to parasitoids}

\item{\code{introduction_date}}{Date parasitoids are introduced in YYYY-MM-DD format}

\item{\code{introduction_n}}{Number of adult parasitoids introduced (assumed to all be female)}
}}

\examples{
Wasp <- new("parasitoid",
            species = "Diaeretiella rapae",
            fun_dev_para = EndoSim:::fit_lactin(0.13048, 7.64313, -0.01308, 35.47),
            fun_para_scal = EndoSim:::fit_null(1),
            fun_attack = EndoSim:::fit_thirdpoly(0.00001084, -0.0021, -0.0272, -0.0145),
            fun_handling = EndoSim:::fit_thirdpoly(0.298, 0.0035, 0.000256, 1.395, inv = T),
            susc_stage = c(2, 3),
            introduction_date = "2024-04-17",
            introduction_n = 5)

Wasp
}
