% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_infect.R
\name{plant_infect}
\alias{plant_infect}
\title{Plant infections}
\usage{
plant_infect(Ph, Pi, Awt, Ai, Awtp, Aip, Ac, temperature, prop_day, Amt)
}
\arguments{
\item{Ph}{Population of healthy plants (Plants/m2)}

\item{Pi}{Population of infected plants (Plants/m2)}

\item{Awt}{Density of wild type insects (Insects/m2)}

\item{Ai}{Density of infected insects (Insects/m2)}

\item{Awtp}{Number of newly produced wild type insects in time step (Insects/m2)}

\item{Aip}{Number of newly produced infected insects in time step (Insects/m2)}

\item{Ac}{Carrying capacity of paddock (Insects/m2)}

\item{temperature}{Mean temperature during time step (C)}

\item{prop_day}{Proportion of day within temperature thresholds}

\item{Amt}{Movements of insects per day}
}
\value{
\code{Pi} the new population of infected plants at time step
}
\description{
Calculate number of infected plants per time step
}
\details{
Calculate the new population of infected plants at the end of the time step, using the following set of equations:
\deqn{
A_{f} = \frac{{A_{wt} + A_{i}}}{{A_{c}}}
}

\deqn{
P_{n} = \frac{{A_{mt} \cdot A_{i}}}{{(A_{wt} + A_{i}) \cdot A_{n}}}
}

\deqn{
P_{i} = P_{h} \cdot \left(1 - e^{-\frac{{P_{n}}}{{P_{h} + P_{i}}}}\right)
}
}
\keyword{internal}
