% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,sim_conds,missing-method}
\alias{plot,sim_conds,missing,ANY-method}
\alias{plot,endosim_mod,missing-method}
\alias{plot,endosim_mod,missing,ANY-method}
\alias{plot,endosim_col,missing-method}
\alias{plot,endosim_col,missing,ANY-method}
\alias{plot,pest,missing-method}
\alias{plot,pest,missing,ANY-method}
\title{Plot methods for S4 objects of class sim_conds, endosim_mod, endosim_col, and pest}
\usage{
plot(x, y, ...)

\S4method{plot}{sim_conds,missing}(x, y, ...)

\S4method{plot}{endosim_mod,missing}(x, type = "pop_size", ...)

\S4method{plot}{endosim_col,missing}(x, y, ...)

\S4method{plot}{pest,missing}(x, y, ...)
}
\arguments{
\item{x}{S4 object of class \linkS4class{sim_conds} or \linkS4class{endosim_mod} or \linkS4class{endosim_col} or \linkS4class{pest}}

\item{y}{from the generic \code{plot} function, ignored for EndoSim objects}

\item{...}{Any other argument suitable for \code{plot()}}

\item{type}{"pop_size" (default) to plot population sizes through time, "R+" to plot proportion of R+ through time, "demo" to plot proportion of population by lifestage through time, "para" to plot parasitoid populations through time; ignored for class \link{sim_conds}, \link{endosym_col} and \link{pest}}
}
\description{
Plot methods for S4 objects of class sim_conds, endosim_mod, endosim_col, and pest
}
\keyword{methods}
\keyword{plot}
