% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sampleEndoSim}
\alias{sampleEndoSim}
\title{Create an EndoSim simulation object}
\usage{
sampleEndoSim(
  data,
  sample_size = 100,
  ages = c(30, 40),
  num_prev_losses = 1:5,
  seed = NULL
)
}
\arguments{
\item{data}{A simulation object (required).}

\item{sample_size}{Number of biopsies to retrieve from simulation.}

\item{ages}{Age range for sampling.}

\item{num_prev_losses}{Number of previous losses for sampling.}

\item{seed}{Set seed. Default NULL.}
}
\value{
Table of biopsies
}
\description{
This function initialises an EndoSim object with initial settings. It requires age range and
number of iterations.
}
