% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\docType{methods}
\name{show,pest-method}
\alias{show,pest-method}
\alias{show,pest,ANY-method}
\alias{show,endosym-method}
\alias{show,endosym,ANY-method}
\alias{show,crop-method}
\alias{show,crop,ANY-method}
\alias{show,parasitoid-method}
\alias{show,parasitoid,ANY-method}
\alias{show,endosim_mod-method}
\alias{show,endosim_mod,ANY-method}
\title{Show methods for S4 objects of class pest, endosym, crop, parasitoid, and endosim_mod}
\usage{
\S4method{show}{pest}(object)

\S4method{show}{endosym}(object)

\S4method{show}{crop}(object)

\S4method{show}{parasitoid}(object)

\S4method{show}{endosim_mod}(object)
}
\arguments{
\item{object}{S4 object of class \linkS4class{pest}, \linkS4class{endosym}, \linkS4class{crop}, \linkS4class{parasitoid}, or \linkS4class{endosim_mod}}
}
\description{
Show methods for S4 objects of class pest, endosym, crop, parasitoid, and endosim_mod
}
\keyword{methods}
\keyword{show}
