% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_conds.R
\docType{class}
\name{sim_conds-class}
\alias{sim_conds-class}
\alias{sim_conds}
\title{An S4 class to store simulation conditions for the endosymbiont model}
\description{
An S4 class to store simulation conditions for the endosymbiont model
}
\section{Slots}{

\describe{
\item{\code{env}}{dataframe containing temperature and precipitation per day of simulation}

\item{\code{sim_length}}{length of simulation in days}

\item{\code{start_date}}{first day of simulation in YYYY-MM-DD format}
}}

