% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\docType{class}
\name{species-class}
\alias{species-class}
\alias{species}
\title{An S4 class to store a species' life history and functions for the endosymbiont model}
\description{
An S4 class to store a species' life history and functions for the endosymbiont model
}
\section{Slots}{

\describe{
\item{\code{name}}{The name of the species}

\item{\code{reproduction}}{The form of reproduction the species has ("sexual", "asexual")}

\item{\code{development}}{The form of development of the species ("holometabolous", "hemimetabolous")}

\item{\code{f_growth}}{Function for growth rate ~ temperature}

\item{\code{f_loss}}{Function for mortality rate ~ rainfall}

\item{\code{f_suscept}}{Function for infection susceptability ~ temperature}

\item{\code{f_inoc}}{Function for inoculation efficiency ~ temperature}

\item{\code{f_alate}}{Function for production of alates ~ proportion of carrying capacity (density)}
}}

\examples{
GPA <-
new("species",
  name = "Myzus persicae",
  reproduction = "asexual",
  development = "hemi",
  f_growth = EndosymbiontModel:::fit_tpc("custom",
    c(4, 8.87, 11.82, 14.22, 17.13, 19.92, 22.61, 25.77, 27.80, 30),
    c(0, 0.06, 0.09, 0.11, 0.13, 0.15, 0.16, 0.14, 0.04, 0)),
  f_loss = EndosymbiontModel:::fit_sigmoid(0.55, 1.5, 3.5),
  f_suscept = EndosymbiontModel:::fit_tpc("custom",
    seq(0, 40, 4)[1:10],
    c(0, 0.15, 0.3, 0.45, 0.6, 0.75, 0.75, 0.6, 0.3, 0)),
 f_inoc = EndosymbiontModel:::fit_tpc("quadratic", 0.8, 22, 0, 40),
 f_alate = EndosymbiontModel:::fit_tpc("custom",
  seq(0, 1, 0.1),
  c(0.01, 0.22, 0.35, 0.45, 0.53, 0.6, 0.66, 0.72, 0.76, 0.78, 0.79))
 )

GPA
}
