% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-methods.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,endosim_col-method}
\alias{summary,endosim_col,missing,ANY-method}
\title{Summary method for S4 object of class endosim_col}
\usage{
summary(x, ...)

\S4method{summary}{endosim_col}(x, ...)
}
\arguments{
\item{x}{S4 object of class \code{endosim_col}}

\item{...}{Any other argument suitable for summary()}
}
\description{
Summary method for S4 object of class endosim_col
}
\keyword{methods}
\keyword{summary}
