% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEndosimObject.R
\name{createEndoSim}
\alias{createEndoSim}
\title{Create an EndoSim simulation object}
\usage{
createEndoSim(iter_num = 1000, age_range = 20:45)
}
\arguments{
\item{iter_num}{Number of iterations of the simulation}

\item{age_range}{Range of ages}
}
\value{
An EndoSim object
}
\description{
This function initialises an EndoSim object with initial settings. It requires age range and
kepted.
}
