% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_custom.R
\name{fit_custom}
\alias{fit_custom}
\title{Fit custom curve}
\usage{
fit_custom(x, y, k = NULL)
}
\arguments{
\item{x}{vector of numeric predictor variable}

\item{y}{vector of numeric response variable, of same length as x}

\item{k}{passed onto smooth term of x. Default is length(x) - 1}
}
\value{
function that accepts x and returns y based on fitted model
}
\description{
Fits a curve using a generalised additive model
}
\details{
Fits a generalised additive model (GAM) using the mgcv package.
}
\keyword{internal}
