% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gaussian.R
\name{fit_gaussian}
\alias{fit_gaussian}
\title{Fit Gaussian curve}
\usage{
fit_gaussian(ymax, xopt, xmin, xmax)
}
\arguments{
\item{ymax}{maximal value of y}

\item{xopt}{value of x at which y = ymax}

\item{xmin}{minimal threshold of x (y = 0)}

\item{xmax}{maximal threshold of x (y = 0)}
}
\value{
function that accepts x and returns y based on fitted model
}
\description{
Fits a curve using the Gaussian model for fitting thermal performance curves
}
\details{
Fits a curve using the Gaussian model:
\deqn{
  y = y_{\text{max}} \cdot e^{-0.5 \cdot \left(\frac{{|x - x_{\text{opt}}|}}{{a}}\right)^2}
}
Optimal value for \code{a} is estimated based on provided arguments.
}
\examples{
my_func <- EndosymbiontModel:::fit_gaussian(0.16, 22.61, 4, 30)
plot(seq(0, 50, by = 0.1), my_func(seq(0, 50, by = 0.1)), type = "l")
}
\seealso{
\code{\link[=fit_briere]{fit_briere()}}, \code{\link[=fit_custom]{fit_custom()}}, \code{\link[=fit_sigmoid]{fit_sigmoid()}}, \code{\link[=fit_quadratic]{fit_quadratic()}}, \code{\link[=fit_rezende]{fit_rezende()}}, \code{\link[=fit_weibull]{fit_weibull()}}
}
\keyword{internal}
