% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_quadratic.R
\name{fit_quadratic}
\alias{fit_quadratic}
\title{Fit Brière curve}
\usage{
fit_quadratic(ymax, xopt, xmin, xmax)
}
\arguments{
\item{ymax}{maximal value of y}

\item{xopt}{value of x at which y = ymax}

\item{xmin}{minimal threshold of x (y = 0)}

\item{xmax}{maximal threshold of x (y = 0)}
}
\value{
function that accepts x and returns y based on fitted model
}
\description{
Fits a curve using the Brière model for fitting thermal performance curves
}
\details{
Fits a curve using the quadratic model:
\deqn{
  y = a + b \cdot x + c \cdot x^2
}
Optimal values for \code{a} (parameter defining y when x = 0), \code{b}, and \code{c} are estimated based on provided arguments.
}
\keyword{internal}
