% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sigmoid.R
\name{fit_sigmoid}
\alias{fit_sigmoid}
\title{Fit sigmoidal curve}
\usage{
fit_sigmoid(K, B, M)
}
\arguments{
\item{K}{upper asymptote}

\item{B}{growth rate}

\item{M}{time of maximum growth}
}
\value{
function that accepts x and returns y based on fitted model
}
\description{
Fits a sigmoidal growth function that returns value based on day after emergence
}
\details{
Fits a sigmoidal growth function to produce a value ranging from 0 to K:
\deqn{
  \frac{K}{1 + e^{-B \cdot \left(x - M \right)}}
}
}
\examples{
my_func <- EndoSim:::fit_sigmoid(1, 0.1, 10)
plot(seq(0, 200), my_func(seq(0, 200)), type = "l")
}
\seealso{
\code{\link[=fit_custom]{fit_custom()}}
}
\keyword{internal}
