% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_weibull.R
\name{fit_weibull}
\alias{fit_weibull}
\title{Fit Weibull curve}
\usage{
fit_weibull(ymax, xopt, xmin, xmax)
}
\arguments{
\item{ymax}{maximal value of y}

\item{xopt}{value of x at which y = ymax}

\item{xmin}{minimal threshold of x (y = 0)}

\item{xmax}{maximal threshold of x (y = 0)}
}
\value{
function that accepts x and returns y based on fitted model
}
\description{
Fits a curve using the Weibull model for fitting thermal performance curves
}
\details{
Fits a curve using the Weibull model:
\deqn{
  y = a \cdot \left(\frac{c - 1}{c}\right)^{\frac{1 - c}{c}} \cdot \left(\frac{x - x_{\text{opt}}}{b} + \left(\frac{c - 1}{c}\right)^{\frac{1}{c}}\right)^{c - 1}  \cdot e^{-\left(\left(\frac{x - x_{\text{opt}}}{b} + \frac{c - 1}{c}\right)^{\frac{1}{c}}\right)^c + \frac{c - 1}{c}\right)} 
}
Optimal values for \code{a} (curve height parameter), \code{b} (curve breadth parameter), and \code{c} (curve shape parameter) are estimated based on provided arguments.
}
\keyword{internal}
