% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial.R
\docType{class}
\name{initial-class}
\alias{initial-class}
\title{An S4 class to store initial population sizes for the endosymbiont model}
\description{
An S4 class to store initial population sizes for the endosymbiont model
}
\section{Slots}{

\describe{
\item{\code{Pest}}{Array describing sizes of pest cohorts of each developmental stage, divided into R+ and R- apterae and alates}

\item{\code{Crop}}{Numeric vector of length 3 describing the crop. First element is proportion of R- plants (should usually be 1), second element is proportion of R+ plants (should usually be 0), third element is number of plants}
}}

