% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pest.R
\docType{class}
\name{pest-class}
\alias{pest-class}
\title{An S4 class to store a pest's life history for the endosymbiont model}
\description{
An S4 class to store a pest's life history for the endosymbiont model
}
\section{Slots}{

\describe{
\item{\code{species}}{The name of the pest}

\item{\code{bg_loss}}{Background mortality (0-1)}

\item{\code{alate_penalty}}{Relative fecundity penalty of alates compaered to apterae (0-1)}

\item{\code{apterae_walk}}{Number of plants per day visited by apterae}

\item{\code{alate_flight}}{Number of plants per day visited by alates}

\item{\code{fun_dev_apt}}{Function describing temperature-dependent development of nymphs destined to become apterae}

\item{\code{fun_dev_ala}}{Function describing temperature-dependent development of nymphs destined to become alates}

\item{\code{fun_imi_neg}}{Function describing number of immigrant R- alates}

\item{\code{fun_imi_pos}}{Function describing number of immigrant R+ alates}

\item{\code{fun_emi}}{Function describing number of emigrants among newly metamorphosed alates}

\item{\code{fun_temp_loss}}{Function describing daily mortality due to exceeding temperature thresholds}

\item{\code{fun_rainfall_loss}}{Function describing daily mortality due to rainfall}

\item{\code{fun_sen_loss}}{Function describing daily mortality due to senescence}

\item{\code{fun_dens_fecund}}{Function describing daily fecundity as function of pest density}

\item{\code{fun_temp_fecund}}{Function describing daily fecundity as function of temperature}

\item{\code{fun_age_fecund}}{Function describing daily fecundity as function of adult cohort age}

\item{\code{fun_alate_prod}}{Function describing proportion of newly produced nymphs destined to become alates as function of pest density}
}}

\examples{
GPA <-
new("pest",
  species = "Myzus persicae",
  bg_loss = 0.03,
  alate_penalty = 0.5,
  apterae_walk = 0.0005,
  alate_flight = 0.005,
  fun_dev_apt = EndosymbiontModel:::fit_wang(24.1, 0.152, 18.09, 32.79, 2.84, 1.8977),
  fun_dev_ala = EndosymbiontModel:::fit_wang(24.1, 0.16474, 19.96, 33.17, 1.41, 2.2202),
  fun_imi_neg = EndosymbiontModel:::fit_imi_barton(),
  fun_imi_pos = EndosymbiontModel:::fit_null(0),
  fun_emi = EndosymbiontModel:::fit_emi_thack(),
  fun_temp_loss = EndosymbiontModel:::fit_null(0),
  fun_rainfall_loss = EndosymbiontModel:::fit_rainfall_thack(),
  fun_sen_loss = EndosymbiontModel:::fit_gompertz(0.00188, 0.13),
  fun_dens_fecund = EndosymbiontModel:::fit_bannerman(10000, 0.0008),
  fun_temp_fecund = EndosymbiontModel:::fit_quadratic(-0.7611, 31.9847, 25),
  fun_age_fecund = EndosymbiontModel:::fit_weibull(58, 1.885, 5.953, 0),
  fun_alate_prod = EndosymbiontModel:::fit_alate(67.418, 0.993, 0.076, 300)
 )

GPA
}
