% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sampleEndoSim}
\alias{sampleEndoSim}
\title{Sampling an EndoSim simulation object}
\usage{
sampleEndoSim(
  summary_data,
  sample_size = NULL,
  ages = c(30, 40),
  num_prev_losses = 2:5,
  max_cycles = 12,
  seed = NULL
)
}
\arguments{
\item{summary_data}{A simulation summary table (required).}

\item{sample_size}{Number of biopsies to retrieve from simulation. Default NULL.}

\item{ages}{Age range for sampling.}

\item{num_prev_losses}{Number of previous losses for sampling.}

\item{max_cycles}{Maximum number of cycles to outcome}

\item{seed}{Set seed. Default NULL.}
}
\value{
Table of biopsies
}
\description{
This function samples an EndoSim object. It requires age range and
number of iterations.
}
