% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,endosim_mod-method}
\alias{show,endosim_mod,ANY-method}
\title{Show method for S4 object of class endosim_mod}
\usage{
show(object)

\S4method{show}{endosim_mod}(object)
}
\arguments{
\item{object}{object of class \linkS4class{endosim_mod}}
}
\description{
Show method for S4 object of class endosim_mod
}
\keyword{methods}
\keyword{show}
