% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\docType{methods}
\name{show,endosym_mod-method}
\alias{show,endosym_mod-method}
\alias{show}
\title{Show endosym_mod object}
\usage{
\S4method{show}{endosym_mod}(object)
}
\description{
Show endosym_mod object
}
