% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-methods.R
\docType{methods}
\name{summary,endosim_mod-method}
\alias{summary,endosim_mod-method}
\alias{summary,endosim_mod,ANY-method}
\alias{summary,endosim_col-method}
\alias{summary,endosim_col,ANY-method}
\title{Summary method for S4 objects of class endosim_mod or endosim_col}
\usage{
\S4method{summary}{endosim_mod}(object, ...)

\S4method{summary}{endosim_col}(object, ...)
}
\arguments{
\item{object}{S4 object of class \linkS4class{endosim_mod} or \linkS4class{endosim_col}}

\item{...}{Any other argument suitable for summary()}
}
\description{
Summary method for S4 objects of class endosim_mod or endosim_col
}
\keyword{methods}
\keyword{summary}
